Changed the only float parameter parser into a double for homogonies sake.
git-svn-id: http://cscirepo.morris.umn.edu/svn/research/LandscapeEC@10559 e5a9bd0e-96e4-4e5d-b9da-31dc9101860d

+++ src/landscapeEC/problem/SeparableProblem.java
@@ -2,5 +2,5 @@ package landscapeEC.problem;
 
 public interface SeparableProblem<T> {
     
-    public abstract SeparableProblem<T> crossover(SeparableProblem<T> firstParent, SeparableProblem<T> secondParent, float noiseStrength);
+    public abstract SeparableProblem<T> crossover(SeparableProblem<T> firstParent, SeparableProblem<T> secondParent, double noiseStrength);
 }
